Sir Bilal Khan Assignments/Assignment 04 Custom Web Search Tool Tavily/pyproject.toml
[project]
name = "ai-agent-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "openai-agents>=0.2.9",
    "python-decouple>=3.8",
    "tavily-python>=0.7.11",
]
